+++ src/SurveysFacade.php
@@ -1,0 +1,16 @@
+<?php 
+
+namespace Sirs\Surveys;
+
+use Illuminate\Support\Facades\Facade;
+
+/**
+* SUrveys facade
+*/
+class SurveysFacade extends Facade
+{
+    protected static function getFacadeAccessor()
+    {
+        return 'surveys';
+    }
+}