+++ src/SurveysServiceProvider.php
@@ -26,8 +26,9 @@ class SurveysServiceProvider extends ServiceProvider
     {
         $this->loadViewsFrom(__DIR__.'/Views', 'surveys');
 
+        $this->loadMigrationsFrom(__DIR__.'/database/migrations/');
+
         $this->publishes([ __DIR__.'/config/config.php' => config_path('surveys.php') ], 'config');
-        $this->publishes([ __DIR__.'/database/migrations/' => database_path('/migrations') ], 'migrations');
         $this->publishes([ __DIR__.'/../../../assets/sass/'=> base_path('/resources/assets/sass')], 'sass');
         $this->publishes([ __DIR__.'/../../../assets/js/'=> public_path('/js')], 'js');
         $this->publishes([ __DIR__.'/Policies/'=> app_path('Policies')], 'policies');